+++ src/components/footer.js
@@ -1,6 +1,6 @@
 import React from "react"
 import { CONFIG } from "./config.js"
-// import Icon from './icons/icon.js'; // Corrected import statement
+import Icon from "./icons/icon.js"
 
 const Footer = () => {
   const showFooter = CONFIG.showFooter
@@ -17,11 +17,7 @@ const Footer = () => {
             >
               <i className="fa-regular fa-envelope"></i>
             </a>
-            <a
-              href="https://nightowllabs.io"
-              target="_blank"
-              rel="noreferrer"
-            >
+            <a href="https://nightowllabs.io" target="_blank" rel="noreferrer">
               <i className="fa-regular fa-building"></i>
             </a>
             <a
@@ -66,6 +62,40 @@ const Footer = () => {
             >
               <i className="fa-solid fa-palette"></i>
             </a>
+            <a
+              href="https://marketplace.visualstudio.com/publishers/scottgriv"
+              target="_blank"
+              rel="noreferrer"
+            >
+              <Icon name="VSCode" />
+            </a>
+          </div>
+        </div>
+        <div
+          style={{
+            display: "flex",
+            justifyContent: "center", // Center horizontally
+            alignItems: "center", // Center vertically
+            padding: '10px',
+          }}
+        >
+          <div
+            style={{
+              height: "32px",
+              width: "114px",
+              overflow: "hidden",
+              borderRadius: "6px",
+            }}
+          >
+            <iframe
+              src="https://github.com/sponsors/scottgriv/button"
+              title="Sponsor scottgriv"
+              style={{
+                border: 0,
+                height: "100%",
+                width: "100%",
+              }}
+            />
           </div>
         </div>
         <div id="pagefoot">