+++ .github/CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2024-06-14
+
+### Changed
+
+- Updated footer links.
+
 ## [1.0.1] - 2024-06-13
 
 ### Added
 
-- Visual Studio Marketplace [URL](https://marketplace.visualstudio.com/items?itemName=scottgriv.Dark-Castle)
+- Visual Studio Marketplace [URL](https://marketplace.visualstudio.com/items?itemName=scottgriv.Dark-Castle).
 
 ## [1.0.0] - 2024-06-10
 
@@ -17,5 +23,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Official release of **Dark Castle Website**.
 
+[1.0.2]: https://github.com/scottgriv/Dark-Castle-Website/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/scottgriv/Dark-Castle-Website/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/scottgriv/Dark-Castle-Website/releases/tag/v1.0.0